pyproject.toml
[project]
name = "agentic-ai-workspace"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "certifi>=2025.7.14",
    "grandalf>=0.8",
    "ipykernel>=6.29.5",
    "ipython>=9.4.0",
    "langchain>=0.3.26",
    "langchain-core>=0.3.69",
    "langchain-groq>=0.3.6",
    "langchain-mcp-adapters>=0.1.9",
    "langchain-tavily>=0.2.10",
    "langgraph>=0.5.3",
    "langsmith>=0.4.7",
    "mcp>=1.12.2",
    "pip>=25.1.1",
    "python-dotenv>=1.1.1",
]
